backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.12
alembic==1.10.4
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.6
bcrypt==4.0.1
httpx==0.24.0
pytest==7.3.1
asyncpg==0.27.0
email-validator==2.0.0
openai==0.27.6
tenacity==8.2.2
python-dateutil==2.8.2
